Update minimum CMake to 2.8.12
The latest version of CMake gives a deprecation warning if you
set the minimum required CMake to earlier than that.

Technically, the minimum version of CMake is not defined by
UseLATEX.cmake itself, but it is defined in all the test and
in an the main documentation build, which itself is used as
an example.