Be less clever about README links
Previously, the download links in the README file contained the
full URL to the latest version of the CMake and PDF. The problem
is that this URL is fragile. It always points to the top copy
(as opposed to the current commit if looking at a particular
verion). Plus the URL could change with updates to GitLab or
other configuration changes.

Just link to the file and let GitLab provide the right page.
This is a bit less nice as you get the GitLab source view, but
you still get a download link right on it. I'm sure users can
figure it out.